Add C collation and indices to Party and ServiceResource columns (#445)
Adds b-tree indices to ServiceOwner and Party columns, which greatly
improves query performance on the enduser search endpoint. `ANALYZE
"Dialog"` should be performed to enable the query planner to properly
utilize the indices.

This also reworks the local development implementation of Altinn
Authorization, causing it to not explode on non-trivial amounts of data
in the database.