chore: Enable K6 functional test run after app deploy to test/staging (#643)
This enables automatic running of K6 E2E tests after successful app
revision deploy

## Description

This refactors the dispatch-only K6 runner to a reusable action, which
is used in CI/CD workflows as well as a dispatch-workflow, which now
takes a path to a test suite to run, eliminating the need for different
dispatches for functional and non-functional tests.

Reporting of the tests are performed with a third-party action-plugin to
utilize GIthub support for junit.xml format. Having some issues getting
this to work properly locally with `act`, so this will have to be tested
in Github. Unclear how rich the reporting within Github can be, but
hopefully we can at least get the job to be marked as failed if any
tests fail as reported by K6. For additional details, the stdout log
will have to be consulted.

In the future, we might want to incorporate more custom handling of the
summary and eg. integrate the Slack Notifier bot.

## Related Issue(s)

- #80 

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)