chore(azure): fix property conflict for redis (#826)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

The default property seems to be publicNetworkAccess: 'Disabled', not
possible to both configure private link and disable public network
access at the same time it seems. Encountered this issue:
https://github.com/Azure/azure-cli/issues/20997

Also, seems like we need to actually create the private DNS Zone
manually, although the documentation states otherwise. Other sources say
that if we use Bicep for this, we need to create it manually :)

Ran the Bicep manually from local machine, seems to update just fine. 

## Related Issue(s)

- #{issue number}

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)