feat: Add doc blocks on DTOs for OAS generation, CORS headers (#987)
## Description

This adds doc blocks for the root aggregate entities, which is a
sufficient start for the documentation generator on Altinn Docs. The
DTOs belonging to "sub" endpoints (transmissions, activities, seen logs)
and parties are not (yet) updated (will be mostly duplicated texts).

In addition, some cleanups have been done in the DTOs:
- Id properties for actors have been removed from DTOs. These are not
mutable, and the id has therefor no reason to be exposed (as opposed to
eg. actions and attachments, as well as the immutable lists in which the
id is useful for idempotency handling)
- Id properties for all end user entities except dialogs themselves and
those available at sub-endpoints and need to be individually adressable
(transmissions, activity log entries, seen log entries) have been
removed. The end user API consumers has no real use for this
information, so we remove it.
- createdAt and updatedAt has been added to the CreateDialogDto used in
the service owner API. Both these fields must be possible to provide
upon dialog creation. **Note!** UpdatedAt handling does not work, as the
IUpdateable entity handling will always override.
- removed Id from CreateDialogDialogAttachment and
CreateDialogTransmissionAttachment. This was a remnant from before
DialogElements was renamed to DialogAttachments. As with eg. actions,
the service owner cannot dictate the id for these entities (but can use
those generated by Dialogporten as idempotency checks in updates)

## Related Issue(s)

- #85 

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [x] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [x] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)

---------

Co-authored-by: Ole Jørgen Skogstad <skogstad@softis.net>
Co-authored-by: Magnus Sandgren <5285192+MagnusSandgren@users.noreply.github.com>
Co-authored-by: Are Almaas <arealmaas@gmail.com>