ci: add production workflow (#1019)
<!--- Provide a general summary of your changes in the Title above -->

## Description
Will run a dry-run of the deployment to both infra and apps on a new
release. Will have to be reviewed manually before deployment happens.

We now have a full workflow for deployment to production. Some time in
the not so very near future, I think it would be wise to split up the
dry-run part and the deployment itself. As right now it would require a
review from any of us to even start the dry-run, which should happen
automatically on every release.
<!--- Describe your changes in detail -->

## Related Issue(s)

- #845 

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)