fix: Add separate settings for parties cache, don't cache invalid response from Altinn 2 (#1194)
## Description

Currently, if the user is missing a profile in Altinn 2, the reportee
list returned from all auth/am APIs are empty. This is an error
condition, so we treat it as such and don't cache the empty response.

Also, this introduces a separate cache setting for parties, as this does
not have the cardinality issues as the PDP cache, allowing us to utilize
memory cache for this.

## Related Issue(s)

- N/A

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced caching mechanism for authorized parties, improving
performance and response times.
- Introduced a new exception handling mechanism for upstream service
errors.

- **Bug Fixes**
- Improved error handling in the `UpstreamServiceException` class to
allow custom messages.

- **Documentation**
- Updated configuration parameters for caching authorized parties to
ensure clarity on retention and timeout settings.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->