feat(infrastructure): add availability test for apim (#1327)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

An availability test for the backend. For now will send a health-check
request to web-api-so so verify that the service is up and running with
the all dependencies. The health-endpoint in APIM will send requests to
web-api-so by default, the other services are not exposed yet.

- Adds an availability test towards our APIM. Will probe the deep
version of the health-checks which checks third party URLs together with
Redis and Postgres.
- Will now only target web-api-so as it is the default backend. Should
expose all services like this.

<img width="612" alt="image"
src="https://github.com/user-attachments/assets/a368ed4d-78c5-4966-b363-493c85bd4568">

The frontend availability test:


![image](https://github.com/user-attachments/assets/55cbe387-d246-4b45-bbd4-17722f4117ab)

## Related Issue(s)

- #292 

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced a new parameter `apimUrl` for capturing the APIM instance
URL across various environments (production, staging, test, yt01).
- Added a new module for creating an availability test for the APIM
instance, enhancing monitoring capabilities.

- **Enhancements**
- New output declaration for the Application Insights resource ID,
allowing easier access to the resource identifier post-deployment.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->