feat: Synchronization of resource policy metadata (#1411)
## Description

This adds a command akin to sync-subject-resources which fetches
information about resource policies and stores the information to a
table in postgres. For now, this only parses and stores XACML
obligations for required authentication level, but can be fairly easily
extended to implement eg. #40

There are however some real limitations to the Resource Registry as of
now for Altinn Apps, as they are not really stored there, but merely
have a representation that does not include the policy. ~~For these, we
store a default authentication level.~~ Edit: We don't do this, as we
cannot safely assume anything here. So we do not store anything unless
we have an explicit obligation for authentication level set in a policy
we can reach. For all other resources, no authentication level handling
will be performed.

## Related Issue(s)

- #1214 
 
## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Release Notes

- **New Features**
	- Introduced a new command for synchronizing resource policies.
- Added support for managing resource policy information in various
environments (production, staging, test).
- Enhanced deployment capabilities with a new scheduled job for resource
policy synchronization.
  
- **Bug Fixes**
	- Improved error handling and logging during synchronization processes.

- **Documentation**
- Updated command documentation to include new synchronization options
and parameters.

- **Chores**
- Added new settings to local development configurations to manage
synchronization behavior at startup.
- Introduced additional configuration settings for local development to
control synchronization processes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Magnus Sandgren <5285192+MagnusSandgren@users.noreply.github.com>
Co-authored-by: Ole Jørgen Skogstad <skogstad@softis.net>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>