fix: Collapse subject resource mappings before building sql query (#1579)
## Description

This implements a few optimizations for search authorization, which
particularly improves performance for users with access to a large
number of parties.
- SubjectsByParties gets collapsed and merged into ResourcesByParties.
This lets us avoid doing subselects in the OR-clauses added to the
search SQL
- The entire SubjectResources-table is cached, which lets us do the
above without an additional database hit.
- When constructing the SQL, parties with identical resource lists are
grouped, further collapsing multiple OR clauses into single ones with
two ANY checks. This vastly reduces the number of clauses for large
number of reportees, which in turn reduces the number of (identical)
query parameters being sent over the network.
- Minor improvements to data structures (mostly replacing List<T> with
HashSet<T>)

## Related Issue(s)

- N/A

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [x] Relevant automated test added