fix(azure): adjust SKU and storage for staging (#1601)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

Related to
https://digdir.slack.com/archives/C079D6PAGDS/p1734018202302459

Experiencing heavy CPU usage on the postgresql server in staging. After
some investigations, it was hard to pin down the cause. Being on a
burstable tier may be too fragile, and perhaps contribute to the high
CPU usage, so upgrading. Staging will be used by many test environments
moving forward so making it more stable with GeneralPurpose machines
makes sense.

## Related Issue(s)

- #N/A

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)