feat(apps): add otel exporter for graphql, service and web-api (#1528)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

- Sets up local OTEL setup that match the OTEL configuration in Azure
Container Apps
- Added fusioncache telemetry
- Added Entity Framework telemetry
- Missing liveMetrics if we want that. Needs to be considered. Other
than that, the most relevant traces are pulled out from the AzureMonitor
package.
- Metrics are only visible locally for now. Turns out that the Azure
Monitor Workspace has a Prometheus instance, but it does not allow for
us sending metrics to it, as it does not have an OTEL endpoint 🙃
Solution here was adding the MetricsMonitor to send metrics directly to
app insights for now..!
- Will add logging in the next PR

To see your metrics, spin up the OTEL services by running
`docker-compose-otel.yml`. The service should start sending to the OTLP
collector automatically.

Example of a trace in the local Jaeger:

![CleanShot 2024-12-09 at 17 52
11@2x](https://github.com/user-attachments/assets/295eba27-84e8-4735-9a0e-be4f2fcfed9c)

## Related Issue(s)

- #1465 

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)

---------

Co-authored-by: Ole Jørgen Skogstad <skogstad@softis.net>