fix(deps): ensure traces are sent to application insights (#1776)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

After some debugging, we found out that the only thing different that is
related to opentelemetry are updates of these dependencies.

1.47.1 seemed to work, but not 1.47.2.
https://github.com/Altinn/dialogporten/compare/88a9217846afaa1580f9430c52e2f880bd4906c1...d9281fc697b6ad613c5546b1cd81a115e349bde4

(this is a bit fragile, so we need to consider other options, or adding
alerts that notify us if we don't get any traces anymore)

Reverted commit: 75c7a24f3897eb6a64dc63463d3db492f44ebd79.

## Related Issue(s)

- #{issue number}

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)