chore: refactor production deployment flow (#1771)
## Description

<!--- Describe your changes in detail -->

Separate dry-run of production deployments and actual deployment to
production. Using the same repository dispatch as for yt01 and staging
when release is created, but instead of deploying, will run a dry-run.

We have an issue here where we would checkout `main` when running
deployment to production, meaning we get the bicep configuration which
is in `main` and not in the version tag. Will look into a solution for
this in the next PR.

## Related Issue(s)

- #1692

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)