ci: ensure deployments to yt01/staging/prod uses the tag as a base rather than main (#1808)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

- Ensures that deployments to yt01/staging/production does not check out
`main` branch for bicep configuration of an application or infra when a
release is made. There was an issue where a release was created, and
when deploying to yt01, it checked out the `main` branch, causing it to
use bicep config from the `main` branch instead of from the tag.
- Added a required version on the `ci-cd-prod.yml` workflow to avoid
that we check out the `main` branch on deployment
- Removed manual dispatch of the `ci-cd-`-workflows. Keeping these
automatic. We should rather use the `dispatch-apps` and
`dispatch-infrastructure` workflows instead if we want to do a manual
deployment.
- Added some descriptions on the workflows
- For deployments to staging/yt01/prod-dry-run, using the version that
was sent in the repository dispatch instead of reading the current
version from file.

## Related Issue(s)

- #1692